+++ src/core/__init__.py
@@ -8,14 +8,25 @@ This module provides foundational utilities used across all architectural layers
 The core module has NO dependencies on other application layers.
 """
 
-from src.core.errors import DomainError, ValidationError
+from src.core.errors import (
+    AuthenticationError,
+    AuthorizationError,
+    ConflictError,
+    DomainError,
+    NotFoundError,
+    ValidationError,
+)
 from src.core.enums import ErrorCode
 from src.core.result import Failure, Result, Success
 
 __all__ = [
+    "AuthenticationError",
+    "AuthorizationError",
+    "ConflictError",
     "DomainError",
     "ErrorCode",
     "Failure",
+    "NotFoundError",
     "Result",
     "Success",
     "ValidationError",
