+++ src/infrastructure/errors/infrastructure_error.py
