+++ src/infrastructure/secrets/aws_adapter.py
@@ -7,8 +7,9 @@ File: aws_adapter.py → class AWSAdapter (PEP 8 naming)
 
 import json
 
-from src.core.errors import ErrorCode, SecretsError
+from src.core.enums import ErrorCode
 from src.core.result import Failure, Result, Success
+from src.domain.errors import SecretsError
 
 
 class AWSAdapter:
