+++ tests/unit/test_infrastructure_secrets_env_adapter.py
@@ -19,7 +19,8 @@ from unittest.mock import patch
 
 import pytest
 
-from src.core.errors import ErrorCode, SecretsError
+from src.core.enums import ErrorCode
+from src.domain.errors import SecretsError
 from src.core.result import Failure, Success
 from src.infrastructure.secrets.env_adapter import EnvAdapter
